refactor(append): rename database_inserter files to table_inserter
