+++ performance/many_string_equals.cpp
@@ -9,8 +9,7 @@
 #include <fmt/format.h>
 
 #include "sequence_generator.h"
-#include "silo/append/database_inserter.h"
-#include "silo/append/ndjson_line_reader.h"
+#include "silo/append/table_inserter.h"
 #include "silo/initialize/initializer.h"
 #include "silo/query_engine/exec_node/ndjson_sink.h"
 #include "silo/query_engine/operators/aggregate_node.h"
