+++ performance/mutation_benchmark.cpp
@@ -5,8 +5,7 @@
 #include <utility>
 
 #include "sequence_generator.h"
-#include "silo/append/database_inserter.h"
-#include "silo/append/ndjson_line_reader.h"
+#include "silo/append/table_inserter.h"
 #include "silo/initialize/initializer.h"
 #include "silo/query_engine/exec_node/ndjson_sink.h"
 #include "silo/query_engine/planner.h"
